# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Kvarenda rental platform API for Uzbekistan
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: getUsers
      tags: [users]
      summary: Get all users (admin only)
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"

    post:
      operationId: createUser
      tags: [users]
      summary: Create or register a user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserBody"
      responses:
        "201":
          description: Created user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"

    put:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserBody"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
...
  /payments:
    get:
      operationId: getPayments
      tags: [payments]
      summary: Get payment history
      parameters:
        - name: rentalId
          in: query
          schema:
            type: integer
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, completed, failed, declared]
      responses:
        "200":
          description: List of payments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Payment"

    post:
      operationId: createPayment
      tags: [payments]
      summary: Create or process a payment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePaymentBody"
      responses:
        "201":
          description: Created payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /payments/{id}/confirm:
    post:
      operationId: confirmCashPayment
      tags: [payments]
      summary: Owner confirms cash payment receipt
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Confirmed payment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Payment"
  /rentals:
    get:
      operationId: getRentals
      tags: [rentals]
      summary: Get active rentals
      parameters:
        - name: tenantId
          in: query
          schema:
            type: integer
        - name: ownerId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [active, completed, terminated]
      responses:
        "200":
          description: List of rentals
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Rental"
  /rentals/{id}:
    get:
      operationId: getRental
      tags: [rentals]
      summary: Get rental details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Rental details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Rental"
...
components:
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [tenant, owner, admin]
        verified:
          type: boolean
        myIdVerified:
          type: boolean
        avatarUrl:
          type: string
        createdAt:
          type: string
      required: [id, name, email, role, verified, myIdVerified, createdAt]
    CreateUserBody:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        phone:
          type: string
        role:
          type: string
          enum: [tenant, owner, admin]
        avatarUrl:
          type: string
      required: [name, email, role]
    UpdateUserBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        verified:
          type: boolean
        myIdVerified:
          type: boolean
        avatarUrl:
          type: string
...
    Payment:
      type: object
      properties:
        id:
          type: integer
        rentalId:
          type: integer
        tenantId:
          type: integer
        period:
          type: string
        amountUzs:
          type: integer
        serviceFeeUzs:
          type: integer
        totalUzs:
          type: integer
        method:
          type: string
          enum: [online, cash]
        status:
          type: string
          enum: [pending, completed, failed, declared]
        ownerConfirmed:
          type: boolean
        paidAt:
          type: string
        createdAt:
          type: string
      required: [id, rentalId, tenantId, period, amountUzs, serviceFeeUzs, totalUzs, method, status, ownerConfirmed, createdAt]
    CreatePaymentBody:
      type: object
      properties:
        rentalId:
          type: integer
        tenantId:
          type: integer
        period:
          type: string
        amountUzs:
          type: integer
        method:
          type: string
          enum: [online, cash]
      required: [rentalId, tenantId, period, amountUzs, method]
    Rental:
      type: object
      properties:
        id:
          type: integer
        contractId:
          type: integer
        listingId:
          type: integer
        tenantId:
          type: integer
        ownerId:
          type: integer
        tenantName:
          type: string
        ownerName:
          type: string
        listingTitle:
          type: string
        address:
          type: string
        district:
          type: string
        monthlyRentUzs:
          type: integer
        serviceFeeUzs:
          type: integer
        startDate:
          type: string
        endDate:
          type: string
        status:
          type: string
          enum: [active, completed, terminated]
        protectedRent:
          type: boolean
        insuranceStatus:
          type: string
          enum: [none, available, insured]
        nextPaymentDate:
          type: string
        totalPaid:
          type: integer
        paymentsCount:
          type: integer
        photos:
          type: array
          items:
            type: string
        createdAt:
          type: string
      required: [id, contractId, listingId, tenantId, ownerId, monthlyRentUzs, serviceFeeUzs, startDate, endDate, status, protectedRent, createdAt]
...