# lib/api-zod/src/generated/types/rental.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kvarenda rental platform API for Uzbekistan
 * OpenAPI spec version: 0.1.0
 */
import type { RentalInsuranceStatus } from "./rentalInsuranceStatus";
import type { RentalStatus } from "./rentalStatus";

export interface Rental {
  id: number;
  contractId: number;
  listingId: number;
  tenantId: number;
  ownerId: number;
  tenantName?: string;
  ownerName?: string;
  listingTitle?: string;
  address?: string;
  district?: string;
  monthlyRentUzs: number;
  serviceFeeUzs: number;
  startDate: string;
  endDate: string;
  status: RentalStatus;
  protectedRent: boolean;
  insuranceStatus?: RentalInsuranceStatus;
  nextPaymentDate?: string;
  totalPaid?: number;
  paymentsCount?: number;
  photos?: string[];
  createdAt: string;
}