# lib/api-zod/src/generated/types/user.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Kvarenda rental platform API for Uzbekistan
 * OpenAPI spec version: 0.1.0
 */
import type { UserRole } from "./userRole";

export interface User {
  id: number;
  name: string;
  email: string;
  phone?: string;
  role: UserRole;
  verified: boolean;
  myIdVerified: boolean;
  avatarUrl?: string;
  createdAt: string;
}